Update README.md
For the course on R we had some weekly assignments. I uploaded the assignments for the first week. We also had to conduct a final project: we had to clean a nested dataset in a group and then develop a research question and answer it through some visualisation. I also uploaded this project. 
For the course on Python that I am now attending, I uploaded the first exam that we did as well as my first attempt at web scraping.