Add files via upload
This is the second exam that I had to do for the course 'Big Data & Automated Content Analysis'. In this exam, I performed automated content analysis. After choosing the best classifier and tuning it, I trained it on a dataset containing already coded news articles. Then, I used it to predict unlabeled news articles to determine which one was about politics.